+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - [Config] Allow to store the config file in "/etc/lithops/config"
+- [CLI] Alow to specify 'memory' and 'version' in the 'lithops runtime delete' command
 
 ### Changed
 - [Core] Include function result in the status.json file if it is < 8KB